+++ src/Laravel/AppPool.php
@@ -3,6 +3,7 @@
 namespace HuangYi\Shadowfax\Laravel;
 
 use HuangYi\Shadowfax\Contracts\AppPool as PoolContract;
+use HuangYi\Shadowfax\Events\AppPoppedEvent;
 use HuangYi\Shadowfax\Events\AppPushingEvent;
 use HuangYi\Shadowfax\HasEventDispatcher;
 use Illuminate\Contracts\Container\Container;
@@ -47,6 +48,8 @@ class AppPool implements PoolContract
      */
     public function pop(): Container
     {
+        $this->dispatch(AppPoppedEvent::class, $this->app);
+
         return $this->app;
     }